// packages/contracts/src/enums.ts
export enum AgentRole {
  DIRECTOR = 'DIRECTOR',
  DEVELOPER = 'DEVELOPER',
  DESIGNER = 'DESIGNER',
  CONTENT = 'CONTENT',
  TESTER = 'TESTER',
}

export enum AgentStatus {
  ENTERING = 'ENTERING',
  WALKING_TO_DESK = 'WALKING_TO_DESK',
  WORKING = 'WORKING',
  WALKING_AROUND = 'WALKING_AROUND',
  RETURNING_TO_DESK = 'RETURNING_TO_DESK',
}